+++ src/parseTools.js
@@ -1477,31 +1477,34 @@ function makeStaticString(string) {
   return '(stringToUTF8("' + string + '", ' + ptr + ', ' + len + '), ' + ptr + ')';
 }
 
-// Generates access to module exports variable in pthreads worker.js
+// Generates access to module exports variable in pthreads worker.js. Depending on whether main code is built with MODULARIZE
+// or not, asm module exports need to either be accessed via a local exports object obtained from instantiating the module (in src/worker.js), or via
+// the global Module exports object.
 function makeAsmExportAccessInPthread(variable) {
   if (MODULARIZE) {
-    return "Module['" + variable + "']" // 'Module' is defined in worker.js local scope, so not EXPORT_NAME in this case.
+    return "Module['" + variable + "']"; // 'Module' is defined in worker.js local scope, so not EXPORT_NAME in this case.
   } else {
-    return EXPORT_NAME + "['" + variable + "']"
+    return EXPORT_NAME + "['" + variable + "']";
   }
 }
 
-// Generates access to a global scope variable in pthreads worker.js
+// Generates access to a JS global scope variable in pthreads worker.js. In MODULARIZE mode the JS scope is not directly accessible, so all the relevant variables
+// are exported via Module. In non-MODULARIZE mode, we can directly access the variables in global scope.
 function makeAsmGlobalAccessInPthread(variable) {
   if (MODULARIZE) {
-    return "Module['" + variable + "']" // 'Module' is defined in worker.js local scope, so not EXPORT_NAME in this case.
+    return "Module['" + variable + "']"; // 'Module' is defined in worker.js local scope, so not EXPORT_NAME in this case.
   } else {
-    return variable
+    return variable;
   }
 }
 
 // Generates access to both global scope variable and exported Module variable, e.g. "Module['foo'] = foo" or just plain "foo" depending on if we are MODULARIZEing.
-// Used the be able to initialize both variables at the same time.
-function makeAsmExportAndGlobalAccessInPthread(variable) {
+// Used the be able to initialize both variables at the same time in scenarios where a variable exists in both global scope and in Module.
+function makeAsmExportAndGlobalAssignTargetInPthread(variable) {
   if (MODULARIZE) {
-    return "Module['" + variable + "'] = " + variable // 'Module' is defined in worker.js local scope, so not EXPORT_NAME in this case.
+    return "Module['" + variable + "'] = " + variable; // 'Module' is defined in worker.js local scope, so not EXPORT_NAME in this case.
   } else {
-    return variable
+    return variable;
   }
 }
